# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/payment-method": {
            "get": {
                "description": "Get a list of PaymentMethod.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Get all PaymentMethod.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Method"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Method.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Create New Method.",
                "parameters": [
                    {
                        "description": "the body to create a new Method",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.paymentMethodInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Method"
                        }
                    }
                }
            }
        },
        "/payment-method/{id}": {
            "get": {
                "description": "Get an PaymentMethod by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Get PaymentMethod.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Method id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Method"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Method by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Delete one Method.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Method id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Method by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Update Method.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Method id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update payment method",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.paymentMethodInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Method"
                        }
                    }
                }
            }
        },
        "/payment-method/{id}/payments": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get all Payments by MethodId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Get Payments.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Method id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Payment"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.paymentMethodInput": {
            "type": "object",
            "properties": {
                "method_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Method": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "method_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Payment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "method_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}